Updates to support hps-sim and other enhancements

+++ cmake/InstallHpsJava.cmake
@@ -14,30 +14,37 @@ execute_process(COMMAND ${CMAKE_SOURCE_DIR}/scripts/print_hps_java_version.sh ${
 set(HPSJAVA_JAR_NAME     hps-distribution-${HPSJAVA_VERSION}-bin.jar)
 set(HPSJAVA_BIN_JAR      ${HPSJAVA_INSTALL_DIR}/${HPSJAVA_JAR_NAME})
 
-externalproject_add(
+message(STATUS "Using HPS Java tag: ${HPSJAVA_TAG}")
+message(STATUS "Using HPS Java version: ${HPSJAVA_VERSION}")
 
-    ${HPSJAVA}
+if(NOT EXISTS "${HPSJAVA_BIN_JAR}")
 
-    INSTALL_DIR       ${HPSJAVA_INSTALL_DIR}
-    SOURCE_DIR        ${HPSJAVA_BUILD_DIR}
+    externalproject_add(
 
-    GIT_REPOSITORY    "https://github.com/JeffersonLab/hps-java"
-    GIT_TAG           ${HPSJAVA_TAG}
-    GIT_SHALLOW       ON
-    GIT_REMOTE_UPDATE_STRATEGY CHECKOUT
+        ${HPSJAVA}
 
-    UPDATE_COMMAND    cd ${HPSJAVA_BUILD_DIR} && git checkout ${HPSJAVA_TAG} && git pull
-    CONFIGURE_COMMAND ""
-    BUILD_COMMAND     cd ${HPSJAVA_BUILD_DIR} && ${MAVEN_BUILD_COMMAND}
-    INSTALL_COMMAND   mkdir -p ${HPSJAVA_INSTALL_DIR} && cp ${HPSJAVA_BUILD_DIR}/distribution/target/${HPSJAVA_JAR_NAME} ${HPSJAVA_INSTALL_DIR}
-)
+        INSTALL_DIR       ${HPSJAVA_INSTALL_DIR}
+        SOURCE_DIR        ${HPSJAVA_BUILD_DIR}
 
-add_dependencies(external ${HPSJAVA})
+        GIT_REPOSITORY    "https://github.com/JeffersonLab/hps-java"
+        GIT_TAG           ${HPSJAVA_TAG}
+        GIT_SHALLOW       ON
+        GIT_REMOTE_UPDATE_STRATEGY CHECKOUT
+
+        UPDATE_COMMAND    cd ${HPSJAVA_BUILD_DIR} && git checkout ${HPSJAVA_TAG} && git pull
+        CONFIGURE_COMMAND ""
+        BUILD_COMMAND     cd ${HPSJAVA_BUILD_DIR} && ${MAVEN_BUILD_COMMAND}
+        INSTALL_COMMAND   mkdir -p ${HPSJAVA_INSTALL_DIR} && cp ${HPSJAVA_BUILD_DIR}/distribution/target/${HPSJAVA_JAR_NAME} ${HPSJAVA_INSTALL_DIR}
+    )
+
+    add_dependencies(external ${HPSJAVA})
+
+    message(STATUS "${HPSJAVA} bin jar will be installed to: ${HPSJAVA_BIN_JAR}")
+
+else()
+    message(STATUS "${HPSJAVA} will not be rebuilt. The bin jar is already installed at: ${HPSJAVA_BIN_JAR}")
+endif()
 
 install(DIRECTORY ${HPSJAVA_BUILD_DIR}/detector-data/detectors DESTINATION ${CMAKE_INSTALL_PREFIX}/share
         FILES_MATCHING PATTERN "*.lcdd"
         PATTERN "SamplingFractions" EXCLUDE)
-
-message(STATUS "Using HPS Java tag: ${HPSJAVA_TAG}")
-message(STATUS "Using HPS Java version: ${HPSJAVA_VERSION}")
-message(STATUS "${HPSJAVA} bin jar will be installed to: ${HPSJAVA_BIN_JAR}")
